+++ lib/src/utils/debouncer.dart
@@ -16,7 +16,7 @@ class Debouncer {
 
   Debouncer({this.milliseconds = 300});
 
-  run(VoidCallback action) {
+  void run(VoidCallback action) {
     _timer?.cancel();
 
     _timer = Timer(Duration(milliseconds: milliseconds), action);