+++ lib/src/widgets/marquee_widget.dart
@@ -23,7 +23,7 @@ class Marquee extends StatelessWidget {
 
   final ScrollController scrollController = ScrollController();
 
-  scroll() async {
+  Future<void> scroll() async {
     while (true) {
       if (scrollController.hasClients) {
         await Future.delayed(pauseDuration);